workaround asynchronous tty bugs on Node.js
